Get Offset
Get a offset array from an input array. The offset array has a padding parameter that should be the size of the patterns that will be read. The offset array is used so that even the cells at the corners can be part of all the possible pattern combinations. To avoid seams.