Inverted Axis
Inverted axis on the function responsible for instantiating the tile objects in the display. Depending on the complexity of the input, it may present a no found neighbor error. As a temporary fix you may double the length of the input array by copying it and pasting it as a seamless square.